added one publisher for the velocity and two subs for start and pwmduty cycle.

+++ lib/ros_lib/time.cpp
@@ -1,4 +1,4 @@
-/* 
+/*
  * Software License Agreement (BSD License)
  *
  * Copyright (c) 2011, Willow Garage, Inc.
@@ -32,19 +32,18 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include "ros.h"
 #include "ros/time.h"
 
 namespace ros
 {
-  void normalizeSecNSec(unsigned long& sec, unsigned long& nsec){
-    unsigned long nsec_part= nsec % 1000000000UL;
-    unsigned long sec_part = nsec / 1000000000UL;
+  void normalizeSecNSec(uint32_t& sec, uint32_t& nsec){
+    uint32_t nsec_part= nsec % 1000000000UL;
+    uint32_t sec_part = nsec / 1000000000UL;
     sec += sec_part;
     nsec = nsec_part;
   }
 
-  Time& Time::fromNSec(long t)
+  Time& Time::fromNSec(int32_t t)
   {
     sec = t / 1000000000;
     nsec = t % 1000000000;
@@ -57,7 +56,7 @@ namespace ros
     sec += rhs.sec;
     nsec += rhs.nsec;
     normalizeSecNSec(sec, nsec);
-    return *this; 
+    return *this;
   }
 
   Time& Time::operator -=(const Duration &rhs){
@@ -66,5 +65,4 @@ namespace ros
     normalizeSecNSec(sec, nsec);
     return *this;
   }
-
 }
